agents/crewai-flows/requirements.txt
# CrewAI Optimization Flow Requirements

# Core dependencies
crewai>=0.11.0
crewai-tools>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# HTTP client for MCP
requests>=2.31.0

# Data validation
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
black>=23.0.0
ruff>=0.1.0
